requirements.txt
kivy
kivymd
requests
pillow
sqlite3
